ustawiony time delay na synapsach (wspolrzedne dodane tez do TypePolls - nie wiadomo czy przydatne)

+++ trunk/NeuralNetwork/src/neuronPackage/Neuron.java
@@ -12,7 +12,7 @@ public class Neuron implements NetworkNode {
 	private int neuronId;
 	private final ArrayList<Synapse> neuronConnections = new ArrayList<Synapse>();
 
-	private final double[] coordinates = new double[3];
+	private final int[] coordinates = new int[3];
 
 	public Neuron(double[] parameters, Type type) {
 		v = -65;
@@ -33,13 +33,13 @@ public class Neuron implements NetworkNode {
 
 	}
 
-	public void setCoordinates(double[] coor) {
+	public void setCoordinates(int[] coor) {
 		for (int i = 0; i < 3; i++) {
 			coordinates[i] = coor[i];
 		}
 	}
 
-	public double[] getCoordinates() {
+	public int[] getCoordinates() {
 		return coordinates;
 	}
